+++ CHANGELOG.md
@@ -13,3 +13,4 @@
 - Added sign() to sign a message or transaction and get signature along with v,r,s.
 - Added getBalance() to fetch the balance in native currency.
 - Added getFee() method for gas estimation for Legacy transactions.
+- Added support for Legacy transactions.